Added code to retrieve cars from the database and and be logged in the html

+++ src/main/java/aaa/pfa/carAuctionBackend/services/CarUploadDTO.java
@@ -9,7 +9,7 @@ public record CarUploadDTO(
         @NotBlank String make,
         @NotBlank String model,
         @NotBlank int year,
-        @NotBlank int price,
+        @NotBlank double price,
         @NotBlank Long id
 ) {
